requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.0
django-grappelli==4.0.1
django-import-export==4.2.1
django-jazzmin==3.0.0
et-xmlfile==1.1.0
openpyxl==3.1.5
pillow==11.0.0
sqlparse==0.5.1
tablib==3.7.0
tzdata==2024.1
